+++ Tests/AccessibilityTests/AccessibilityProject/Sources/TargetA/PublicWrappedProperty.swift
@@ -1,0 +1,17 @@
+import Foundation
+
+@propertyWrapper public struct PublicWrapper {
+  public var wrappedValue: String
+
+  public init(wrappedValue: String) {
+    self.wrappedValue = wrappedValue
+  }
+}
+
+public struct PublicWrappedProperty {
+  @PublicWrapper public var wrappedProperty: String
+
+  public init() {
+      wrappedProperty = ""
+  }
+}