+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - A newline is no longer printed before non-Xcode formatted results.
 - `--external-codable-protocols` now retains enums that conform to `CodingKey`.
 - Fix redundant public accessibility analysis false-positive for actors. 
+- Fix redundant public accessibility analysis false-positive for property wrappers.
 
 ## 2.17.1 (2023-12-04)